+++ Quaver.API/Maps/Processors/Difficulty/Rulesets/Keys/DifficultyConstantsKeys.cs
@@ -17,40 +17,40 @@ namespace Quaver.API.Maps.Processors.Difficulty.Rulesets.Keys
         /// <summary>
         ///     Base Difficulty Multiplier
         /// </summary>
-        public float DifficultyMultiplier { get; } = 11.2f;
+        public double DifficultyMultiplier { get; } = 11.2f;
 
         /// <summary>
         ///     Base Offset Value. Difficulty will be subtracted by this value.
         ///     - Lowest Difficulty possible in HandState is 1. 
         /// </summary>
-        public float DifficultyOffset { get; } = -24f;
+        public double DifficultyOffset { get; } = -24f;
 
         /// <summary>
         ///     Determines if the HitObject on the same hand of a specific HitObject will be counted as a Chorded Hit Object.
         /// </summary>
-        public float ChordThresholdSameHandMs { get; } = 8f;
+        public double ChordThresholdSameHandMs { get; } = 8f;
 
         /// <summary>
         ///     Determines if the HitObject on the other hand of this HitObject will be counted as a Chorded Hit Object.
         /// </summary>
-        public float ChordThresholdOtherHandMs { get; } = 16f;
+        public double ChordThresholdOtherHandMs { get; } = 16f;
 
         /// <summary>
         ///     Any two states with intervals below this value will be considered vibro.
         ///     - a 1/4 170bpm jack has an interval of 88.235ms
         /// </summary>
-        public float VibroActionToleranceMs { get; } = 88.2f;
+        public double VibroActionToleranceMs { get; } = 88.2f;
 
         /// <summary>
         ///     Any two states with intervals below this value will fall under the Vibro gradient.
         ///     - a 1/4 150bpm jack has an interval of 100ms
         /// </summary>
-        public float VibroActionThresholdMs { get; } = 100;
+        public double VibroActionThresholdMs { get; } = 100;
 
         /// <summary>
         /// Max Absolute Delta between two Simple Jack Actions to be considered as repetition.
         /// </summary>
-        public float WristGapDeltaThresholdMs { get; } = 26;
+        public double WristGapDeltaThresholdMs { get; } = 26;
 
         /// <summary>
         ///     Max Repetition for Simple Jack. Will not count simple jacks that repeat after this.
@@ -60,12 +60,12 @@ namespace Quaver.API.Maps.Processors.Difficulty.Rulesets.Keys
         /// <summary>
         ///     Max Delta between two objects for stamina calculation.
         /// </summary>
-        public float MaxStaminaDelta { get; } = 1000;
+        public double MaxStaminaDelta { get; } = 1000;
 
         /// <summary>
         ///     Determines the Multiplier for 1/4th Beat Action Length
         /// </summary>
-        public float BpmToActionLengthMs { get; } = 15000;
+        public double BpmToActionLengthMs { get; } = 15000;
 
         /// <summary>
         ///     Spline Multiplier for Hand States that have higher Difficulty than the previous predecessor
